requirements.txt
alpaca-py>=0.18.0
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
tzdata>=2024.1
